fix(models): song_id in history data type int from 32 to 64 bits

+++ server/app/history/models.py
@@ -4,5 +4,5 @@ from users.models import Users
 # Create your models here.
 class History(models.Model):
     user = models.ForeignKey(Users, on_delete=models.CASCADE)
-    song_id = models.PositiveIntegerField()
+    song_id = models.PositiveBigIntegerField()
     order  = models.PositiveSmallIntegerField()
